+++ src/imports/completions/skillfileCompletionProvider.ts
@@ -4,7 +4,7 @@ import { mechanicsDataset, targetersDataset, conditionsDataset, ObjectType, Obje
 import { getAllAttributes, getMechanicDataByName } from '../utils/mechanicutils';
 import { getObjectLinkedToAttribute } from '../utils/cursorutils';
 import { isEnabled } from '../utils/configutils';
-import { isMetaskillFile } from '../utils/configutils';
+import { isMetaskillFile, enableDashesSuggestions } from '../utils/configutils';
 
 export const SkillFileCompletionProvider = vscode.languages.registerCompletionItemProvider(
     'yaml',
@@ -25,12 +25,23 @@ export const SkillFileCompletionProvider = vscode.languages.registerCompletionIt
             const indentpre = document.lineAt(position.line - 1).firstNonWhitespaceCharacterIndex;
             let indentation = " ".repeat(indentpre - indentnow);
 
+            // if (context.triggerCharacter != undefined) {
+            //     const previousLine = document.lineAt(position.line - 1).text;
+            //     const shouldDeletePreviousLine = ["-",""].includes(previousLine.trim());
+            //     if (shouldDeletePreviousLine) {
+            //         const range = new vscode.Range(position.line - 1, 0, position.line - 1, previousLine.length);
+            //         const edit = new vscode.WorkspaceEdit();
+            //         edit.replace(document.uri, range, " ".repeat(previousLine.length));
+            //         await vscode.workspace.applyEdit(edit);
+            //         return undefined;
+            //     }    
+            // }
 
             if(keys.length == 0){
                 return undefined;
             }
             else if(keys.length == 2){
-                if(SkillFileObjects[keys[0] as keyof typeof SkillFileObjects].type == "list"){
+                if(enableDashesSuggestions() && SkillFileObjects[keys[0] as keyof typeof SkillFileObjects].type == "list"){
                     const completionItem = new vscode.CompletionItem("-", vscode.CompletionItemKind.Snippet);
                     completionItem.insertText = new vscode.SnippetString(indentation + "- $0");
                     completionItem.command = { command: 'editor.action.triggerSuggest', title: 'Re-trigger completions...' };